+++ crates/macro/ui-tests/async-errors.rs
@@ -1,3 +1,4 @@
+#![allow(unreachable_code)]
 use wasm_bindgen::prelude::*;
 
 #[wasm_bindgen]