+++ tests/wasm/imports.rs
@@ -51,6 +51,12 @@ extern "C" {
     fn unused_import();
     fn assert_dead_import_not_generated();
     fn should_call_undefined_functions() -> bool;
+
+    type StaticMethodCheck;
+    #[wasm_bindgen(static_method_of = StaticMethodCheck)]
+    fn static_method_of_right_this();
+
+    static STATIC_STRING: String;
 }
 
 #[wasm_bindgen]
@@ -231,4 +237,14 @@ fn undefined_function_is_ok() {
     let x = TypeThatIsNotDefined::new();
     x.method();
     x.set_property(x.property());
+}
+
+#[wasm_bindgen_test]
+fn static_string_ok() {
+    assert_eq!(*STATIC_STRING, "x");
+}
+
+#[wasm_bindgen_test]
+fn static_method_of_has_right_this() {
+    StaticMethodCheck::static_method_of_right_this();
 }